backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
pypdf
python-dotenv
ddgs
tenacityhttpx
sqlalchemy
ddgs
pydantic
streamlit
beautifulsoup4
